backend/llm/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv

langchain
langgraph
langsmith

langchain-openai
langchain-community
openai
requests

# 테스트 관련
pytest
pytest-asyncio
pytest-mock

# 캐싱 (선택사항)
redis

# 로깅
structlog

# 유효성 검증
email-validator

# 코드 품질 및 포맷팅
flake8
black
isort
mypy
pre-commit
pylint

# 타입 스텁
types-requests
